# Make SetF faster by not using variables