After a little testing, it seems that user variables aren't the fastest compared to repeating expressions:

```
adamj@localhost [3]> select benchmark(1000000, (select if(first_name > '', concat_ws('', first_name, ''), '') from actor limit 1));
+--------------------------------------------------------------------------------------------------------+
| benchmark(1000000, (select if(first_name > '', concat_ws('', first_name, ''), '') from actor limit 1)) |
+--------------------------------------------------------------------------------------------------------+
|                                                                                                      0 |
+--------------------------------------------------------------------------------------------------------+
1 row in set (0.01 sec)

adamj@localhost [4]> select benchmark(1000000, (select if((@f:=first_name) > '', concat_ws('', @f, ''), '') from actor limit 1));
+------------------------------------------------------------------------------------------------------+
| benchmark(1000000, (select if((@f:=first_name) > '', concat_ws('', @f, ''), '') from actor limit 1)) |
+------------------------------------------------------------------------------------------------------+
|                                                                                                    0 |
+------------------------------------------------------------------------------------------------------+
1 row in set (1.56 sec)
```

And to prove the duplicates aren't being optimized away (though that wouldn't necessarily be bad):

```
adamj@localhost [6]> explain extended select if(first_name > '', concat_ws('', first_name, ''), '') from actor limit 1;
+------+-------------+-------+------+---------------+------+---------+------+------+----------+-------+
| id   | select_type | table | type | possible_keys | key  | key_len | ref  | rows | filtered | Extra |
+------+-------------+-------+------+---------------+------+---------+------+------+----------+-------+
|    1 | SIMPLE      | actor | ALL  | NULL          | NULL | NULL    | NULL |  200 |   100.00 |       |
+------+-------------+-------+------+---------------+------+---------+------+------+----------+-------+
1 row in set, 1 warning (0.00 sec)

adamj@localhost [7]> show warnings
    -> ;
+-------+------+---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| Level | Code | Message                                                                                                                                                                                     |
+-------+------+---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| Note  | 1003 | select if((`sakila`.`actor`.`first_name` > ''),concat_ws('',`sakila`.`actor`.`first_name`,''),'') AS `if(first_name > '', concat_ws('', first_name, ''), '')` from `sakila`.`actor` limit 1 |
+-------+------+---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
1 row in set (0.00 sec)

adamj@localhost [8]> explain extended select if((@f:=first_name) > '', concat_ws('', @f, ''), '') from actor limit 1;
+------+-------------+-------+------+---------------+------+---------+------+------+----------+-------+
| id   | select_type | table | type | possible_keys | key  | key_len | ref  | rows | filtered | Extra |
+------+-------------+-------+------+---------------+------+---------+------+------+----------+-------+
|    1 | SIMPLE      | actor | ALL  | NULL          | NULL | NULL    | NULL |  200 |   100.00 |       |
+------+-------------+-------+------+---------------+------+---------+------+------+----------+-------+
1 row in set, 1 warning (0.00 sec)

adamj@localhost [9]> show warnings;
+-------+------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| Level | Code | Message                                                                                                                                                                  |
+-------+------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| Note  | 1003 | select if(((@f:=`sakila`.`actor`.`first_name`) > ''),concat_ws('',(@`f`),''),'') AS `if((@f:=first_name) > '', concat_ws('', @f, ''), '')` from `sakila`.`actor` limit 1 |
+-------+------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
1 row in set (0.00 sec)
```