@adamchainz 2015-04-16
Actually looks like this is a side-effect of the way BENCHMARK works in MySQL, nevermind:

```
In [19]: %timeit cursor.execute("select if(first_name > '', concat_ws('', first_name, ''), '') from actor limit 1")
10000 loops, best of 3: 85.9 µs per loop

In [20]: %timeit cursor.execute("select if((@f:=first_name) > '', concat_ws('', @f, ''), '') from actor limit 1")
10000 loops, best of 3: 86.2 µs per loop
```